modules/health_chatbot/requirements.txt
# modules/health_chatbot/requirements.txt
langchain==0.2.17
flask==3.0.3
sentence-transformers==3.2.1
pypdf==5.4.0
python-dotenv==1.0.1
langchain-pinecone==0.1.3
langchain-openai==0.1.0
langchain-community==0.2.19
pinecone-client==5.0.1
ipykernel==6.29.5